mitsui-commodity-prediction/requirements.txt
# Data processing
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1

# Machine Learning
scikit-learn==1.3.0
lightgbm==4.0.0
xgboost==1.7.6
catboost==1.2

# Deep Learning (optional)
torch==2.0.1
tensorflow==2.13.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Jupyter
jupyterlab==4.0.3
ipywidgets==8.1.0
notebook==7.0.0

# Utilities
tqdm==4.65.0
pyyaml==6.0.1
python-dotenv==1.0.0

# Hyperparameter optimization
optuna==3.2.0

# Experiment tracking (optional)
mlflow==2.5.0
wandb==0.15.8

# Time series
statsmodels==0.14.0
prophet==1.1.4

# Feature engineering
feature-engine==1.6.1
tsfresh==0.20.1

# Model interpretability
shap==0.42.1
lime==0.2.0.1

# Data validation
great-expectations==0.17.12
pandera==0.16.1